backend/requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
django-environ
django-filter
openai
django-cors-headers